Rename the directories for rekor-cli and rekor-server.
This makes the binaries "go installable" by their canonical names.
