+++ cmd/rekor-cli/app/format/wrap.go
