+++ cmd/rekor-cli/app/get.go
@@ -24,7 +24,7 @@ import (
 	"time"
 
 	"github.com/go-openapi/runtime"
-	"github.com/sigstore/rekor/cmd/cli/app/format"
+	"github.com/sigstore/rekor/cmd/rekor-cli/app/format"
 	"github.com/sigstore/rekor/pkg/generated/client/entries"
 	"github.com/sigstore/rekor/pkg/generated/models"
 	"github.com/sigstore/rekor/pkg/log"
