+++ cmd/rekor-cli/app/pflags_test.go
