+++ cmd/rekor-cli/app/root.go
