+++ cmd/rekor-cli/app/root_test.go
