+++ cmd/rekor-cli/main.go
@@ -13,10 +13,9 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-
 package main
 
-import "github.com/sigstore/rekor/cmd/server/app"
+import "github.com/sigstore/rekor/cmd/rekor-cli/app"
 
 func main() {
 	app.Execute()
