+++ cmd/rekor-server/main.go
@@ -13,9 +13,10 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
 package main
 
-import "github.com/sigstore/rekor/cmd/cli/app"
+import "github.com/sigstore/rekor/cmd/rekor-server/app"
 
 func main() {
 	app.Execute()
